bin/darkirc/script/bots/requirements_telegram-mirror-bot.txt
python-telegram-bot==21.9
